+++ apiv4/user_routes.py
@@ -85,11 +85,6 @@ def account_details():
     return response
 
 
-@app.route('/v4/users/gcp/validate/<gcp_id>/', methods=['GET'], strict_slashes=False)
-def validate_gcp_old(gcp_id):
-    return redirect(url_for('validate_gcp', gcp_id=gcp_id), HTTP_301_MOVED_PERMANENTLY)
-
-
 @app.route('/v4/users/cloud_projects/validate/<gcp_id>/', methods=['GET'], strict_slashes=False)
 def validate_gcp(gcp_id):
     """
@@ -137,11 +132,6 @@ def validate_gcp(gcp_id):
     response.status_code = code
         
     return response
-
-
-@app.route('/v4/users/gcp/<gcp_id>/', methods=['DELETE', 'PATCH', 'GET'], strict_slashes=False)
-def user_gcp_old(gcp_id):
-    return redirect(url_for('user_gcp', gcp_id=gcp_id), HTTP_301_MOVED_PERMANENTLY)
 
 
 @app.route('/v4/users/cloud_projects/<gcp_id>/', methods=['DELETE', 'PATCH', 'GET'], strict_slashes=False)
@@ -236,11 +226,6 @@ def user_gcp(gcp_id):
     response.status_code = code
 
     return response
-
-
-@app.route('/v4/users/gcp/', methods=['POST', 'GET'], strict_slashes=False)
-def user_gcps_old():
-    return redirect(url_for('user_gcps'), HTTP_301_MOVED_PERMANENTLY)
 
 
 @app.route('/v4/users/cloud_projects/', methods=['POST', 'GET'], strict_slashes=False)
