+++ gunicorn.conf.py
@@ -1,0 +1,3 @@
+
+# Enforce HTTPS
+secure_scheme_headers = {'X-FORWARDED-PROTOCOL': 'ssl', 'X-FORWARDED-PROTO': 'https', 'X-FORWARDED-SSL': 'on'}
